Add restore paths option to restore script
Need to know which paths rsync should restore to, if we only send the root
of the backup to rsync it will overwrite what is in the restore destination.
So we need to append the first path of the paths to the destination of rsync
and run an rsync command for each path in the config.