Keep the right most slash on the restore path
If the restore path (file or directory) has a parent directory we need to
keep the slash of the restore path. If not, if the restored path is a file
rsync will create a file with the same name as the parent directory with
the content of the file rather than the restored file in the parent directory.