Change rsync shell command
The shell would ignore the quotes when parsing the rsync_port variable
this results in rsync not receiving he full rsh parameter like
--rsh="ssh -p port" instead it gets --rsh="ssh, -p and port", that is
three parameters and when rsync reads the parameter --rsh="ssh it
complains when there is no closing quote. In order to try to be posix
compliant we skip the variable in the script and uses an if branch
instead.