Provide informative exit codes
Blubee continues to backup the items in the 'backup.json' even if there
is an error with some item. This is useful as a partial backup is often
better than no backup. However, when there are alot of items to backup
the error might be hidden in the output and the user may not be aware
that something went wrong. Especially if one is using blubee in a
script, as the final exit code of blubee will be 0. To avoid this we
make sure to detect when ever an item fails to backup and exit with an
approperiate exit code, to notify the caller that something went wrong.

To make it easier to find the error we also provide some logs that may
help the user identify the issue.