Carry exit code to end of execution
When an error happens in the backup script the script will exit with
a non zero exit code. During execution of a 'blubee backup' command the
backup script is run multiple times and the exit code is ignored. This
results in an exit code of 0 from the command even if there were errors
during execution.

This behavior hides the errors in the output and may give the
impression that all files were backed up successfully. We consider this
a fulty behavior.