Create necessary directories even on dry run
During dry runs we don't create any directories. This causes issues for
the rest of the script when the backup target directory doesn't exist.
Rsync expects the parent target directory to exist for it's dry run to
work.

It is important that we don't leave any new files or directories
behind after a dry run so we need to keep track of which branch of the
directory tree we created and remove those directories after the run.

To keep track of what directories we create we need to check what
directories exists both locally and remotely. This means that the ssh
mock needs to be smarter and actually run the requested commands, but do
so locally. Mainly, this is to make sure the "does dir exist?" commands
respond with the happy path.

Since we actually create some directories now we need to also make sure
that tests clean up after them selves (and preferably use unique file
names to avoid clashes.