Use script_name variable in tests
Tests uses a mix between hard coded strings and a script_name variable
at the top of the test file. Many test files prings the wrong script
name. This probably happens beacuse the hard coded script names are at
the bottom of the file and that means they are easy to miss if a test
file is reused for new tests.