Reuse ssh connections to avoid rate limiting
Blubee will sometimes fail intermittently and it seems to sometimes
run into trouble when creating new connections. It is unclear if we are
actually running into rate limiting but it seems we don't get this issue
when using multiplexing (running multiple streams over a single
connection). We configure ssh (both raw and through rsync) to use the
same socket for their connections and thereby avoid alot of auth calls
to the server.

Since it may take a decent amount of time to create a connection this
change also has the benefit that this time is reduces (potentially alot
if the backup config requires a lot of connections).